+++ CHANGES.md
@@ -6,6 +6,7 @@ videojs-record changelog
 
 - New webm-wasm plugin: create webm files using libwebm (compiled with
   WebAssembly) in the browser (#321)
+- Ability to change label of record indicator (#336)
 
 **Backwards-incompatible change** (when upgrading from a previous version):